move funcs to core
